Move old contributions flow selenium tests
